layer(7.2): added Business Connection Class, functions, and updates: Integration with Business Accounts
Added the class BusinessConnection and updates about the connection or disconnection of the bot to a business account, represented by the field business_connection in the class Update.
Added updates about new messages in a business account connected to the bot, represented by the field business_message in the class Update.
Added updates about message edits in a business account connected to the bot, represented by the field edited_business_message in the class Update.
Added updates about message deletion in a business account connected to the bot, represented by the class BusinessMessagesDeleted and the field deleted_business_messages in the class Update.
Added the method getBusinessConnection.